Initial commit
Add the MobileFormatter class and move it into the
Wikimedia\MobileFormatter namespace. Add the MobileFormatterTest test
case and fixtures and move it into the Tests\Wikimedia\MobileFormatter
namespace.

The following changes had to be made in order to get the test passing:

* Use the default value of the MFMobileFormatterHeadings configuration
  variable rather than fetching it via MobileFrontend#getMFConfig
* Use the default value of the MFRemovableClasses configuration variable
  rather than fetching it via MobileFrontend#getMFConfig

These changes are considered to have introduced two bugs - the
pre-existing behaviour could be configured and now can't - but the
latter change is much lower priority as the configuration variable is in
the set of legacy configuration variables.